refactor: 移动类型定义到 napcat_types.py，更新相关导入
